+++ src/backend/joanie/core/api/admin/__init__.py
@@ -754,6 +754,7 @@ class OrderViewSet(
         "certificate",
         "certificate__certificate_definition",
         "credit_card",
+        "voucher",
     )
 
     # Map aliases to actual model field lookups.