+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to
   to purchase a training from an external platform
 - Handle deep link in back office
 - Create standalone to_own orders in admin
+- Expose voucher code and claim status in order admin API
 
 ### Fixed