+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [3.2.2] - 2026-03-20
+
 ### Fixed
 
 - Added missing keys into context for quote arrival request
@@ -812,7 +814,8 @@ and this project adheres to
 - First working version serving sellable micro-credentials for multiple
   organizations synchronized to a remote catalog
 
-[unreleased]: https://github.com/openfun/joanie/compare/v3.2.1...main
+[unreleased]: https://github.com/openfun/joanie/compare/v3.2.2...main
+[3.2.2]: https://github.com/openfun/joanie/compare/v3.2.1...v3.2.2
 [3.2.1]: https://github.com/openfun/joanie/compare/v3.2.0...v3.2.1
 [3.2.0]: https://github.com/openfun/joanie/compare/v3.1.2...v3.2.0
 [3.1.2]: https://github.com/openfun/joanie/compare/v3.1.1...v3.1.2